If /IRQMASK given, auto disable /IRQCHAIN, unless user specifically said to enable both, in which case print warning.
IRQ chaining enabled by default now. It turns out on 2010 or later
motherboards the way to resolve the IRQ storm is to chain to the BIOS
IRQ handler. Whatever way the BIOS does it, is the way to clear the IRQ.

Mask interrupt from IRQ if told to do so. On 2010 Intel i3 it does stop
the IRQ storm but then seems to be unable to restart it again.

Fix IRQ hook/unhook to work even if previous IRQ was NULL.