Either Microsoft found a way to cram an .ICO file into 19 bytes or something is wrong with the rnLength field.
Well, yes. Apparently the rnLength field makes sense if you treat it as
number of 1 << rscAlignShift units, not bytes.

Everyone on the 'net with a copy of the NE spec mis-documenents the
field as bytes. Checking on my copy of the Windows 3.1 SDK, so does
Microsoft. Yet dumping resources from various system and application
files in Windows 3.1 shows rnLength fields that don't make sense unless
adjusted as described above.